Adding error when not run as root

+++ remove_packages.sh
@@ -11,7 +11,14 @@
 # xfonts
 # wbrazilian/catalan/bulgarian/etc
 
-echo "This script should be run as root."
+if [ "$(id -u)" != "0" ]; then
+    echo ""
+    echo ""
+    echo "This script should be run as root."
+    echo ""
+    echo ""
+    exit 1;
+fi
 
 apt-get remove -y                            \
         task-amharic                         \
